+++ whatsonpypi/cli.py
@@ -32,7 +32,33 @@ from .whatsonpypi import get_query_response
     default=False,
     help="Flag to open docs or homepage of project"
 )
-def main(package, more, docs):
+@click.option(
+    '-a',
+    '--add',
+    is_flag=True,
+    required=False,
+    default=False,
+    help="Flag to enable adding of dependencies to requirement files."
+         " By default, it searches for files with names matching requirements*.txt"
+         " in the current working directory and adds the dependency to the end of the"
+         " file. If you want the dependency to be added to a specific line,"
+         " mention a comment '#wopp' on its own line which will be replaced with the dependency."
+         " Existing dependencies will be replaced with newer versions. Dependency version"
+         " by default is the latest unless specified explicitly with 'whatsonpypi package==version'."
+         " Directory to search for requirement files can be specified with --req-path"
+)
+@click.option(
+    '-r',
+    '--req-dir',
+    type=click.Path(exists=True, file_okay=False, dir_okay=True,
+                    readable=True, writable=True, resolve_path=True,
+                    allow_dash=False),
+    required=False,
+    default=".",
+    show_default=True,
+    help="Directory to search for requirement files. Only used when --add is used."
+)
+def main(package, more, docs, add, req_dir):
     """
     CLI tool to find the latest version of a package on PyPI.
 
@@ -51,6 +77,8 @@ def main(package, more, docs):
             version=version,
             more_out=more,
             launch_docs=docs,
+            add_to_req=add,
+            req_dir=req_dir
         )
         # output is not always expected and maybe None sometimes.
         if output: