+++ whatsonpypi/constants.py
@@ -6,3 +6,6 @@ Package level constants
 from __future__ import unicode_literals  # unicode support for py2
 
 PYPI_BASE_URL = 'https://pypi.org/pypi'
+
+REQUIREMENTS_FILE_PATTERN = "requirements*.txt"
+REQUIREMENTS_REPLACE_COMMENT = "#wopp"